remove comm from partial compute+bazel test aligning
