+++ cpp/oneapi/dal/algo/covariance/backend/gpu/partial_compute_kernel.hpp
@@ -18,6 +18,9 @@
 
 #include "oneapi/dal/algo/covariance/compute_types.hpp"
 #include "oneapi/dal/backend/dispatcher.hpp"
+#include "oneapi/dal/backend/primitives/utils.hpp"
+#include "oneapi/dal/util/common.hpp"
+#include "oneapi/dal/detail/policy.hpp"
 
 namespace oneapi::dal::covariance::backend {
