+++ cpp/oneapi/dal/algo/covariance/detail/partial_compute_ops_dpc.cpp
@@ -29,16 +29,14 @@ struct partial_compute_ops_dispatcher<Policy, Float, Method, Task> {
                                             const partial_compute_input<Task>& input) const {
         using kernel_dispatcher_t = dal::backend::kernel_dispatcher< //
             KERNEL_SINGLE_NODE_CPU(backend::partial_compute_kernel_cpu<Float, Method, Task>),
-            KERNEL_UNIVERSAL_SPMD_GPU(backend::partial_compute_kernel_gpu<Float, Method, Task>)>;
+            KERNEL_SINGLE_NODE_GPU(backend::partial_compute_kernel_gpu<Float, Method, Task>)>;
         return kernel_dispatcher_t()(policy, desc, input);
     }
 };
 
-#define INSTANTIATE(F, M, T)                                                        \
-    template struct ONEDAL_EXPORT                                                   \
-        partial_compute_ops_dispatcher<dal::detail::data_parallel_policy, F, M, T>; \
-    template struct ONEDAL_EXPORT                                                   \
-        partial_compute_ops_dispatcher<dal::detail::spmd_data_parallel_policy, F, M, T>;
+#define INSTANTIATE(F, M, T)      \
+    template struct ONEDAL_EXPORT \
+        partial_compute_ops_dispatcher<dal::detail::data_parallel_policy, F, M, T>;
 
 INSTANTIATE(float, method::dense, task::compute)
 INSTANTIATE(double, method::dense, task::compute)
